Update Git aliases and add new commands
This commit updates the Git aliases in the .zshrc file. It also adds new aliases for Git push and pull commands. Additionally, the commit includes changes to the bat alias for better paging and adds the eval command for thefuck and starship. Finally, the commit adds the Volta and Sheldon configurations.